Fix Reddit native Enter key bug
Add a capturing eventListener for 'keydown' events on
SHORTCUT_FOCUSABLE_DIV, which will call event.stopImmediatePropagation()
for when user presses a key, native event for which should be
suppressed. Use it for the 'Enter' key. Currently, for some reason, it
does not work on j/k native shortcuts, so have to call
event.preventDefault() somewhere else (in focusPostOrComment())