Add ShortcutsCategory.initialize()
This function will be called in init.ts when the shortcuts for the
current website are found. Code that should only run for that specific
website will be put inside that function.

Right now, because of how imports are organized (allShortcuts.ts, which
contains shortcuts for all websites is imported into init.ts, which
causes code in all shortcuts/site/category.ts files to be run,
regardless of the current website), this function will allow some code
to only be run for some website.

- Use that function in reddit/posts.ts when disabling native 'Enter'
  shortcut.
- Use that function in google-search/search.ts to call
  updateSearchResults().