Add chrome support to popup and options page
Add a script tag which loads browser-polyfill into those pages. It is
commented out by default and webpack build uncomments it for chromium
build.

Passing message with shortcuts from content script to popup wasn't
working for chrome because the message was not JSON serializable (it
contained Maps). This works in Firefox because it uses the structured
clone algorithm. So had to rewrite it to use regular objects when the
browser is chrome.

Because browser_style is not available in chrome, the popup looks a
little worse than in firefox. Had to adjust padding on span elements
(default style makes letters too wide).